backend/requirements.txt
Flask
Flask-CORS
Flask-Login
Flask-SQLAlchemy
Flask-Mail-SendGrid
Flask-Limiter
SQLAlchemy
PyMySQL
cryptography
Werkzeug
gunicorn
python-dotenv
langchain
langchain-huggingface
openai
huggingface_hub
sentence-transformers
transformers
faiss-cpu
markdown
nltk
urllib3
certifi
numpy
torch
torchvision
torchaudio
